Refactoring code, migrating new classes into net.es.topo.common.dto.esdb.gql package
