+++ backend/src/main/java/net/es/oscars/dto/esdb/gql/GraphqlEsdbEquipmentInterface.java
@@ -1,0 +1,13 @@
+package net.es.oscars.dto.esdb.gql;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+public class GraphqlEsdbEquipmentInterface {
+    private String id;
+    public Integer getId() {
+        return Integer.parseInt(id);
+    }
+}
